+++ src/redirect.js
@@ -3,26 +3,15 @@ import { usePath, getCurrentHash } from './path.js'
 import { navigate } from './navigate.js'
 import { useQueryParams, deriveQueryString } from './querystring.js'
 
-// TODO: V2 replace this signature with one more like the <Redirect /> Component
 export function useRedirect(
   predicateUrl,
   targetUrl,
-  queryParams = null,
-  replace = true
+  { query, replace = true, merge = true } = {}
 ) {
   const currentPath = usePath()
-  useEffect(() => {
-    if (currentPath === predicateUrl) {
-      navigate(targetUrl, queryParams, replace)
-    }
-  }, [predicateUrl, targetUrl, queryParams, replace, currentPath])
-}
-
-export function Redirect({ to, query, replace = true, merge = true }) {
   const [currentQuery] = useQueryParams()
   const hash = getCurrentHash()
-  const path = usePath()
-  let url = to
+  let url = targetUrl
   const targetQuery = deriveQueryString(currentQuery, query, merge)
   if (targetQuery) {
     url += '?' + targetQuery
@@ -30,6 +19,15 @@ export function Redirect({ to, query, replace = true, merge = true }) {
   if (merge && hash && hash.length) {
     url += hash
   }
-  useRedirect(path, url, null, replace)
+
+  useEffect(() => {
+    if (currentPath === predicateUrl) {
+      navigate(url, undefined, replace)
+    }
+  }, [predicateUrl, url, undefined, replace, currentPath])
+}
+
+export function Redirect({ to, query, replace = true, merge = true }) {
+  useRedirect(usePath(), to, { query, replace, merge })
   return null
 }