+++ test/navigate.spec.js
@@ -119,13 +119,14 @@ describe('useNavigate', () => {
     const basePath = '/base'
     const newPath = '/path'
 
-    const { container } = render(
+    const { container, debug } = render(
       <App basePath={basePath} newPath={newPath} replace />
     )
     act(() => navigate(`${basePath}/`))
 
     window.history.replaceState = jest.fn()
     const button = container.querySelector('button')
+    debug()
     act(() => {
       button.dispatchEvent(new MouseEvent('click', { bubbles: true }))
     })